the C file for the project
it includes several functions and algorithms that initiate and start both the lcd and the rtc chip. there are functions responsible of displaying the time and date and the next prayer time in the lcd . and a function to get the current time from the rtc. and a function to compare whether a prayer time has come yet. and a function to enable the alarm after the output of the last function is positive.